# Optimize quicklistCompare with optional string2ll caching.