--- a/src/server.h
+++ b/src/server.h
@@ -2981,7 +2981,7 @@ unsigned char *listTypeGetValue(listTypeEntry *entry, size_t *vlen, long long *l
 void listTypeInsert(listTypeEntry *entry, robj *value, int where);
 void listTypeReplace(listTypeEntry *entry, robj *value);
 int listTypeEqual(listTypeEntry *entry, robj *o, size_t object_len,
-                  long long *cached_val, int *cached_valid);
+                  long long *cached_longval, int *cached_valid);
 void listTypeDelete(listTypeIterator *iter, listTypeEntry *entry);
 robj *listTypeDup(robj *o);
 void listTypeDelRange(robj *o, long start, long stop);