## Optimize quicklistCompare with optional string2ll caching

### Summary

This pull request improves the performance of quicklistCompare and lpCompare by avoiding repeated calls to string2ll when comparing many quicklist/listpack entries against the same string value. The optimization targets use cases like LREM, LPOS, LINSERT, and ZRANK where comparisons are made repeatedly in a loop.

By caching the result of string2ll during a single command execution, we avoid re-parsing the same input string thousands of times—resulting in up to **30% higher throughput and up to 25% lower p50 latency** in LREM LINSERT benchmarks, and **5% higher throughput** in ZRANK (listpack) command.

### Starting point


Profiling the `memtier_benchmark-1key-list-10K-elements-linsert-lrem-integer` benchmark we see string2ll takes 24.3% of the CPU cycles. 

```
(pprof) top
Showing nodes accounting for 22315.79ms, 69.62% of 32052.63ms total
Dropped 674 nodes (cum <= 160.26ms)
Showing top 10 nodes out of 141
      flat  flat%   sum%        cum   cum%
 7789.47ms 24.30% 24.30%  7789.47ms 24.30%  string2ll
 3631.58ms 11.33% 35.63% 10789.47ms 33.66%  quicklistNext
 2789.47ms  8.70% 44.33%  5105.26ms 15.93%  [lto1]
 1736.84ms  5.42% 49.75%  2894.74ms  9.03%  lpValidateNext
 1631.58ms  5.09% 54.84%  5631.58ms 17.57%  lpNext
 1368.42ms  4.27% 59.11%  9473.68ms 29.56%  listTypeEqual
 1052.63ms  3.28% 62.40%  1526.32ms  4.76%  lpGetValue
  947.37ms  2.96% 65.35%  7789.47ms 24.30%  listTypeNext
  789.47ms  2.46% 67.82%   789.47ms  2.46%  lpCurrentEncodedSizeUnsafe (inline)
  578.95ms  1.81% 69.62%   578.95ms  1.81%  lpBytes (inline)
```

### Changes

- Updated quicklistCompare and lpCompare to accept two optional parameters:
  - `long long *cached_val`
  - `int *cached_valid`
- If caching parameters are provided, string2ll is invoked only once and its result is reused across comparisons.
- listTypeEqual was updated to forward these parameters.
- Commands such as LREM, LPOS, LINSERT, and ZRANK now use this optimization.
- All internal tests and usage of quicklistCompare/lpCompare were updated accordingly.

### Behavior

- If cached_valid is NULL, quicklistCompare/lpCompare behaves as before (no caching).
- If cached_valid is non-NULL:
  - 0 means uninitialized: string2ll is attempted.
  - 1 means valid: cached_val is used.
  - -1 means invalid: string2ll previously failed and is skipped.

### Testing

New tests were added to quicklistTest and listpackTest to verify:
- Correctness of comparison with and without caching
- Behavior with numeric and non-numeric strings
- Safety when caching parameters are not used

### Microbenchmark

A benchmark test was added comparing performance with and without caching on a quicklist and listpack. The cached version significantly reduces repeated string2ll calls, showing a clear performance benefit.

#### quicklist

To reproduce it:
```
 make CFLAGS="-DREDIS_TEST" redis-server
./src/redis-server test quicklist --accurate
```

#### quicklistCompare Caching Microbenchmark output

| Mode              | Total Time (usec) | Time per Iteration (usec) | % Change      |
|------------------|------------------:|---------------------------:|--------------:|
| Without Caching  |         7,847,023 |                    156.94 |       —       |
| With Caching     |         5,769,111 |                    115.38 |   −26.5%      |


#### listpack

To reproduce it:
```
 make CFLAGS="-DREDIS_TEST" redis-server
./src/redis-server test listpack --accurate
```

#### quicklistCompare Caching Microbenchmark output

| Mode              | Total Time (usec) | % Change      |
|------------------|------------------:|---------------------------:|--------------:|
| Without Caching  |        28013021 |                    —       |
| With Caching     |        21147508 |        −32.7%      |

### benchmarks using the SPEC

#### LPOS/LINSERT/LREM

```

redis-benchmarks-spec-client-runner --tests-regexp "memtier_benchmark-1key-list-10K-elements-linsert-lrem-integer" --flushall_on_every_test_start --flushall_on_every_test_end  --cpuset_start_pos 2 --override-memtier-test-time 300 --benchmark_local_install --override-test-runs 3 --db_server_port <...> --db_server_password <...> --db_server_host <...>
```

Overall improvement

| Metric             | Baseline (9017871) | Optimized (8f9bda0) | % Change     |
|--------------------|-------------------:|---------------------:|-------------:|
| Ops/sec            |          11469.94  |            14925.09  |    +30.1%    |
| p50 Latency (ms)   |             19.199 |              14.335  |    −25.3%    |
| p99 Latency (ms)   |             23.551 |              19.711  |    −16.3%    |



#### ZRANK

```

redis-benchmarks-spec-client-runner --tests-regexp "memtier_benchmark-1key-zset-listpack-zrank-100-elements-pipeline-1" --flushall_on_every_test_start --flushall_on_every_test_end  --cpuset_start_pos 2 --override-memtier-test-time 300 --benchmark_local_install --override-test-runs 3 --db_server_port <...> --db_server_password <...> --db_server_host <...>
```

Overall improvement

| Metric             | Baseline (117424f85c85cdb9bc560b8fd641ba213534cd50) | Optimized (b2da3c1be5bbd86831bf372bd38c9d20963f6466) | % Change     |
|--------------------|-------------------:|---------------------:|-------------:|
| Ops/sec            |          229771  |            241466  |    +5.1%    |
| p50 Latency (ms)   |             0.735 |              0.711 |    −3.4%    |
| p99 Latency (ms)   |            1.471 |             1.4071  |    −4.5%    |